requirements.txt
discord.py==2.4.0
pokemon_formats==1.0.2
